Calculate canopy height from DSM & DTM, then gets coordinates of all objects
This notebook lets a user choose an ODM task and then subtract the DTM from the DSM to get basic canopy height (CHM). This is then normalised and the locations of all objects in the CHM are calculated and the coordinates are converted to Lat/long so they can be used for additional calculations.
Lat/long data can be pushed to a jupyter notebook that calculates derived satellite indices for all bounding boxes calculated for the CHM.
(foe example using this great GA notebook: https://docs.dea.ga.gov.au/notebooks/Real_world_examples/Vegetation_phenology.html)